Separate Numeric into Integer and Float
Drop dependency on scanf, because it does not currently (ruby 2.5.0) handle
numbers like '1e6' correctly.

+++ test/key_value_name/extension_test.rb
@@ -13,11 +13,11 @@ class TestExtension < MiniTest::Test
     file :foo, :tar, :gz
 
     file :bar, :tar do
-      key :a, type: Numeric
+      key :a, type: Integer
     end
 
     file :bar, :tar, :gz do
-      key :a, type: Numeric
+      key :a, type: Integer
     end
   end
